requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
docutils==0.22.3
filetype==1.2.0
idna==3.11
Kivy==2.3.1
Kivy-Garden==0.1.5
kivymd==1.2.0
pillow==12.0.0
Pygments==2.19.2
requests==2.32.5
speedtest-cli==2.1.3
# Note: speedtest-cli requires a patch for Python 3.13 compatibility
# The patch fixes fileno() issue in Kivy environments
urllib3==2.5.0
